move simhandler
