+++ src/main/java/ecdar/Ecdar.java
@@ -5,12 +5,12 @@ import ecdar.abstractions.Project;
 import ecdar.backend.BackendDriver;
 import ecdar.backend.BackendHelper;
 import ecdar.backend.QueryHandler;
+import ecdar.backend.SimulationHandler;
 import ecdar.code_analysis.CodeAnalysis;
 import ecdar.controllers.EcdarController;
 import ecdar.presentations.BackgroundThreadPresentation;
 import ecdar.presentations.EcdarPresentation;
 import ecdar.presentations.UndoRedoHistoryPresentation;
-import ecdar.simulation.SimulationHandler;
 import ecdar.utility.keyboard.Keybind;
 import ecdar.utility.keyboard.KeyboardTracker;
 import javafx.application.Application;
