+++ src/main/java/ecdar/controllers/TracePaneElementController.java
@@ -3,7 +3,7 @@ package ecdar.controllers;
 import com.jfoenix.controls.JFXRippler;
 import ecdar.Ecdar;
 import ecdar.abstractions.Location;
-import ecdar.simulation.SimulationHandler;
+import ecdar.backend.SimulationHandler;
 import ecdar.simulation.SimulationState;
 import ecdar.presentations.TransitionPresentation;
 import javafx.beans.property.SimpleBooleanProperty;
